pyproject.toml
[project]
name = "agentbench"
version = "0.1.0"
description = "Offline evaluation harness for AI coding agents"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0.0",
    "python-ulid>=2.0.0",
    "path>=17.1.1",
    "filelock>=3.20.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "ruff>=0.1.0"]

[project.scripts]
ab = "agentbench.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentbench"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W"] # Removed "I" (import sorting) - often annoying
ignore = [
    "E501", # Line too long
    "F401", # Unused import (useful during development)
    "F841", # Unused variable (useful during development)
    "E402", # Module level import not at top of file
    "E741", # Ambiguous variable name (l, O, I)
    "W293", # Blank line contains whitespace
]

[tool.pytest.ini_options]
testpaths = ["agentbench"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-m 'not integration and not docker'"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "docker: marks tests as requiring Docker (deselect with '-m \"not docker\"')",
]
